requirements.txt
asgiref==3.5.2
Django==4.1
django-safedelete==1.3.0
sqlparse==0.4.2
psycopg2-binary==2.9.3
factory_boy==3.2.1